Replace paho-mqtt by rumqttc.
rumqttc is a pure Rust MQTT client implementation, where paho-mqtt
depends on the Paho C library (which depends on OpenSSL).

This simplifies the build as we no longer have to static compile
OpenSSL and link against it. It also reduces the binary size by ~ 50%.

+++ src/config.rs
@@ -49,7 +49,7 @@ pub struct Mqtt {
     pub server: String,
     pub username: String,
     pub password: String,
-    pub qos: usize,
+    pub qos: u8,
     pub clean_session: bool,
     pub client_id: String,
     pub ca_cert: String,
